Made friendship its own application
