+++ Wave/friends/admin.py
@@ -1,0 +1,10 @@
+from django.contrib import admin
+
+from .models import *
+
+class FollowAdmin(admin.ModelAdmin):
+    search_fields = ['user1_id', 'user2_id']
+    list_display = ['user1_id', 'user2_id']
+
+
+admin.site.register(Follow, FollowAdmin)
