+++ Wave/users/models.py
@@ -15,17 +15,6 @@ class User(AbstractUser):
     def __str__(self):
         return self.username
 
-class Follow(models.Model):
-
-    user1 = models.ForeignKey(User, related_name="follower", on_delete=models.CASCADE)
-    user2 = models.ForeignKey(User, related_name="followee", on_delete=models.CASCADE)
-
-    class Meta:
-        unique_together = (('user1', 'user2'), )
-
-    def __str__(self):
-        return str(self.id)
-
 
 class Node(models.Model):
